Feat: mute, kick, ban 추가

+++ backend/src/socket/chat/service/chat-user.service.ts
@@ -29,6 +29,10 @@ export class ChatUserService {
       throw new NotAcceptableException(`Password is incorrect`);
     }
 
+    if (chatRoom.bannedUsers.has({ id: userId })) {
+      throw new NotAcceptableException(`User ${userId} is banned in chat room ${chatRoomId}`);
+    }
+
     this.chatRoomRepository.removeSocketFromAllChatRoom(socketId);
     this.chatRoomRepository.addSocketToChatRoom(chatRoomId, socketId, userId);
   }
@@ -44,6 +48,22 @@ export class ChatUserService {
     if (!chatRoom.sockets.has(socketId)) {
       // throw new NotAcceptableException(`Socket ${socketId} is not in chat room ${chatRoomId}`);
       return;
+    }
+
+    if (chatRoom.sockets.size === 1) {
+      this.chatRoomRepository.removeChatRoom(chatRoomId);
+      return;
+    }
+
+    if (chatRoom.sockets.get(socketId).role === 'host') {
+      const nextHost = Array.from(chatRoom.sockets.values()).find(chatUser => chatUser.role === 'operator');
+
+      if (nextHost) {
+        nextHost.role = 'host';
+      } else {
+        const randomHost = Array.from(chatRoom.sockets.values())[0];
+        randomHost.role = 'host';
+      }
     }
 
     this.chatRoomRepository.removeSocketFromChatRoom(chatRoomId, socketId);
@@ -65,14 +85,26 @@ export class ChatUserService {
     return chatRoom;
   }
 
-  public async getChatData(socketId: string, message: string, timestamp: string): Promise<ChatDataDto> {
+  public async createChatData(socketId: string, message: string, timestamp: string): Promise<ChatDataDto> {
     const chatRoom = this.getJoinedChatRoom(socketId);
     const chatUser = chatRoom.sockets.get(socketId);
+
+    if (!chatUser) {
+      throw new NotAcceptableException(`Socket ${socketId} is not in chat room ${chatRoom.id}`);
+    }
+
+    if (chatRoom.bannedUsers.has({ id: chatUser.id })) {
+      throw new NotAcceptableException(`User ${chatUser.id} is banned in chat room ${chatRoom.id}`);
+    }
+
+    if (chatUser.isMutted) {
+      throw new NotAcceptableException(`User ${chatUser.id} is muted in chat room ${chatRoom.id}`);
+    }
 
     return {
       chatUser: {
         user: await this.userRepository.findOneBy({ id: chatUser.id }),
-        isOperator: chatUser.isOperator,
+        role: chatUser.role,
       },
       message,
       timestamp,
